pyproject.toml
[project]
name = "pipeline-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "attrs>=25.3.0",
    "biopython>=1.85",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "polars>=1.28.1",
    "rich>=14.0.0",
    "typer>=0.15.2",
]
